copied and combined the DateService Tests into one file
